requirements.txt
Flask==0.10.1
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-instagram==1.3.2
simplejson==3.8.2
six==1.10.0
Werkzeug==0.11.4
wheel==0.26.0
